Test : GamePad UI 스킬 버튼 드래그 기능 테스트중
스킬 쿨타임중엔 드래그 반응 안하도록 변경

+++ Assets/Scripts/Players_InGame/Player.cs
@@ -67,7 +67,7 @@ public class Player : NetworkBehaviour, IStoreCustomer
             Debug.Log($"spellName: {spellName}");
             ownedSpellPrefabList.Add(GameAssets.instantiate.GetSpellPrefab(spellName));
         }
-        for (int i = 0; i < ownedSpellPrefabList.Count; i++)
+        for (ushort i = 0; i < ownedSpellPrefabList.Count; i++)
         {
             spellController.SetCurrentSpell(ownedSpellPrefabList[i], i);
         }
